Setting up routing structure and update Home component layout
